+++ src/nbody/nbody.cpp
@@ -13,7 +13,7 @@ namespace malloy::nbody
 bool Body2D::is_valid() const
 {
     return mass > math::Real{0} && math::is_finite(mass) &&
-           math::is_finite(position) && math::is_finite(velocity);
+           math::is_squarable(position) && math::is_squarable(velocity);
 }
 
 bool NBodySettings::is_valid() const